+++ src/widgets/Button.h
@@ -13,7 +13,7 @@ QPushButton* create_button_text(const char* icon, ButtonEvent event);
 
 void set_icon(const char* name, QPushButton* button);
 
-void set_shortcut(QPushButton *button, int shortcut, int modifier);
-void do_shortcut(int shortcut, int modifier);
+void set_shortcut(QPushButton *button, qint64 shortcut, qint64 modifier);
+void do_shortcut(qint64 shortcut, qint64 modifier);
 
 #endif // BUTTON_H
